+++ terraform/infra/modules/monitoring/grafana.tf
@@ -195,7 +195,7 @@ resource "aws_ecs_service" "grafana" {
   name            = "cabal-grafana"
   cluster         = var.ecs_cluster_id
   task_definition = aws_ecs_task_definition.grafana.arn
-  desired_count   = 1
+  desired_count   = var.quiesced ? 0 : 1
 
   enable_execute_command = true